+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 
 # v1.1.12
 * Fixed model renderer mode crash
+* Fixed Bloom rendering (thanks to @RogoShum)